requirements.txt
flask==3.1.1
flask-cors==6.0.1
torch==2.8.0
torchvision==0.23.0
transformers==4.36.2
Pillow==10.2.0
numpy==1.26.4
opencv-python==4.12.0.88
easyocr==1.7.2
matplotlib==3.8.2
gunicorn==21.2.0